// Real-Estate-Form/Front/src/app/models/form.ts
export class Form{
    pdfName: string = "";
    name: string = "";
    surname: string = "";
    phone: string = "";
    whatsapp: string = "";
    email: string = "";
    residence: boolean = false;
    visa: boolean = false;
    nationality: string = "";
    moneyQuant: string = "";
    financeType: string = "";
    propertyType: string = "";
    floors: string = "";
    rooms: string = "";
    bathrooms: string =" ";
    garages: string = "";
    constructionState: string = "";
    minConstructionYear: string = "";
    state: string = "";
    city: string = "";
    comunity: string = "";
}